update Ids from int to long

+++ src/Crowdin.Api/StringCorrections/IStringCorrectionsApiExecutor.cs
@@ -6,21 +6,21 @@ namespace Crowdin.Api.StringCorrections
     public interface IStringCorrectionsApiExecutor
     {
         Task<ResponseList<Correction>> ListCorrections(
-            int projectId,
-            int stringId,
+            long projectId,
+            long stringId,
             IEnumerable<SortingRule>? orderBy = null,
             int denormalizePlaceholders = 0,
             int limit = 25,
             int offset = 0);
 
-        Task<Correction> GetCorrection(int projectId, int correctionId);
+        Task<Correction> GetCorrection(long projectId, long correctionId);
         
-        Task<Correction> AddCorrection(int projectId, AddCorrectionRequest request);
+        Task<Correction> AddCorrection(long projectId, AddCorrectionRequest request);
         
-        Task DeleteCorrection(int projectId, int correctionId);
+        Task DeleteCorrection(long projectId, long correctionId);
         
-        Task DeleteCorrections(int projectId, int stringId);
+        Task DeleteCorrections(long projectId, long stringId);
 
-        Task<Correction> RestoreCorrection(int projectId, int correctionId);
+        Task<Correction> RestoreCorrection(long projectId, long correctionId);
     }
 }
